+++ .dockerignore
@@ -9,4 +9,5 @@ README.md
 env.*
 /devops/*
 node_modules
-/mongodb/data
+/data
+doc
