+++ src/messageAppClient/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "messageappclient",
+  "version": "1.0.0",
+  "description": "",
+  "main": "MessageApp.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "axios": "^0.18.0",
+    "debug": "^4.1.0"
+  }
+}
